AllocStrBuf: use size_t as argument type

+++ yamos/YAM_utilities.h
@@ -228,7 +228,7 @@ struct Mail *AddMailToList(struct Mail *mail, struct Folder *folder);
 APTR     AllocCopy(APTR source, int size);
 char *   AllocData2D(struct Data2D *data, LONG initsize);
 char *   AllocReqText(char *s);
-char *   AllocStrBuf(long initlen);
+char *   AllocStrBuf(size_t initlen);
 void     AppendLog(int id, char *text, void *a1, void *a2, void *a3, void *a4);
 void     AppendLogNormal(int id, char *text, void *a1, void *a2, void *a3, void *a4);
 void     AppendLogVerbose(int id, char *text, void *a1, void *a2, void *a3, void *a4);
